refactor: rename @launchstack/search to @launchstack/retrieval
Mechanical rename only — directory, package name, the 13 TS import
sites, five workspace dependents, jest moduleNameMapper, next.config
transpilePackages, the CI publish filter, the e2e-ingest paths, and the
pending changeset. ESLint: the brick list, the per-package block, and
the tools allowlist now say retrieval, and @launchstack/search joins
the legacy-name ban so stale branches fail loudly instead of silently
resolving nothing.
