+++ apps/web/src/lib/tools/rag/retrievers/vector-retriever.ts
@@ -11,4 +11,4 @@ export {
     createDocumentVectorRetriever,
     createCompanyVectorRetriever,
     createMultiDocVectorRetriever,
-} from "@launchstack/search/retrievers";
+} from "@launchstack/retrieval/retrievers";
