+++ packages/engine/package.json
@@ -45,7 +45,7 @@
         "@launchstack/llm": "workspace:^",
         "@launchstack/conversion": "workspace:^",
         "@launchstack/indexing": "workspace:^",
-        "@launchstack/search": "workspace:^",
+        "@launchstack/retrieval": "workspace:^",
         "@launchstack/orchestration": "workspace:^",
         "zod": "^3.23.8"
     },
