+++ packages/retrieval/README.md
@@ -1,18 +1,18 @@
-# @launchstack/search
+# @launchstack/retrieval
 
 Question in, cited answer out: hybrid retrieval (BM25 + vector ensemble behind a replaceable port), second-pass reranking, and the citation builder that turns permission-scoped retrieval rows into stable anchored citations. It deliberately does not contain index-time work — embeddings are generated by llm and persisted by indexing; this package only reads.
 
 ## Install
 
 ```bash
-pnpm add @launchstack/search
+pnpm add @launchstack/retrieval
 ```
 
 ## Use
 
 ```ts
-import { buildCitations } from "@launchstack/search";
-import { configureRag, getRag } from "@launchstack/search";
+import { buildCitations } from "@launchstack/retrieval";
+import { configureRag, getRag } from "@launchstack/retrieval";
 ```
 
 ## API
