+++ packages/retrieval/src/citation-builder.ts
