+++ packages/retrieval/src/index.ts
@@ -1,5 +1,5 @@
 /**
- * @launchstack/search — question in, cited answer out. Hybrid retrieval
+ * @launchstack/retrieval — question in, cited answer out. Hybrid retrieval
  * behind a replaceable port, second-pass reranking, and the citation builder.
  */
 export * from "./hybrid-search";
