+++ packages/tools/package.json
@@ -71,7 +71,7 @@
     "dependencies": {
         "@langchain/core": "^0.3.74",
         "@launchstack/llm": "workspace:^",
-        "@launchstack/search": "workspace:^",
+        "@launchstack/retrieval": "workspace:^",
         "@launchstack/store": "workspace:^",
         "drizzle-orm": "^0.45.1",
         "zod": "^3.23.8"
